Refactor: Opsi task managing
